requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
celery==5.3.4
redis==5.0.1
sqlite-utils==3.35.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
hdbscan==0.8.33
umap-learn==0.5.4
openai==1.3.6
httpx==0.25.2
cryptography==41.0.8 